ConDec-623: Enable more CRUD operations in singleton knowledge graph object for project (#217)
* Remove irrelevant sentences from knowledge graph

* Implement insertDecisionKnowledgeElement in KnowledgePersistenceManager

* Call insertLink and deleteLinks methods of KnowledgePersistenceManager in JiraIssueTextPersistenceManager instead of GenericLinkManager methods

* Make smart linking smarter

* Reset KnowledgeGraph as part of the init method for unit tests

* Remove overwritten equals method in DecisionKnowledgeElement class that prevented the containsVertex method from working

* Try to make updateNode method working, but it is not working yet

* Add .editorconfig file

+++ src/test/java/de/uhd/ifi/se/decision/management/jira/filtering/TestFilterExtractor.java
@@ -24,7 +24,7 @@ public class TestFilterExtractor extends TestSetUp {
 
 	private ApplicationUser user;
 	private FilterSettings settings;
-	
+
 	@Before
 	public void setUp() {
 		init();
@@ -166,7 +166,7 @@ public class TestFilterExtractor extends TestSetUp {
 	@Test
 	@NonTransactional
 	public void testGetAllElementsMatchingCompareFilterSettingsFilledCreated() {
-		settings.setCreatedLatest((long) -1);
+		settings.setCreatedLatest(-1);
 		settings.setCreatedEarliest(System.currentTimeMillis() - 100000);
 		FilterExtractor extractor = new FilterExtractorImpl(user, settings);
 		assertEquals(5, extractor.getAllElementsMatchingCompareFilter().size(), 0.0);
@@ -175,7 +175,7 @@ public class TestFilterExtractor extends TestSetUp {
 	@Test
 	@NonTransactional
 	public void testGetAllElementsMatchingCompareFilterSettingsFilledClosed() {
-		settings.setCreatedEarliest((long) -1);
+		settings.setCreatedEarliest(-1);
 		settings.setCreatedLatest(System.currentTimeMillis() + 1000);
 		FilterExtractor extractor = new FilterExtractorImpl(user, settings);
 		assertEquals(5, extractor.getAllElementsMatchingCompareFilter().size(), 0.0);
